+++ src/core/Renderer.ts
@@ -24,6 +24,20 @@ export class Renderer {
     public markersOnly = false;
 
     /**
+     * When set and {@linkcode naturalSize} is `false` sets the width of the rendered image.
+     * 
+     * Both `width` and `height` have to be set for this to take effect.
+     */
+    public width?: number;
+    /**
+     * When set and {@linkcode naturalSize} is `false` sets the height of the rendered image.
+     * 
+     * Both `width` and `height` have to be set for this to take effect.
+     */
+    public height?: number;
+
+
+    /**
      * Initiates rendering of the result image and returns a promise which when resolved
      * contains a data URL for the rendered image.
      * 
@@ -37,17 +51,40 @@ export class Renderer {
         return new Promise<string>((resolve) => {
             const canvas = document.createElement("canvas");
 
+            const markerImageCopy = document.createElementNS(
+            'http://www.w3.org/2000/svg',
+            'svg'
+            );
+            markerImageCopy.setAttribute('xmlns', 'http://www.w3.org/2000/svg');
+            markerImageCopy.setAttribute('width', markerImage.width.baseVal.valueAsString);
+            markerImageCopy.setAttribute(
+              'height',
+              markerImage.height.baseVal.valueAsString
+            );
+            markerImageCopy.setAttribute(
+              'viewBox',
+              '0 0 ' +
+                markerImage.viewBox.baseVal.width.toString() +
+                ' ' +
+                markerImage.viewBox.baseVal.height.toString()
+            );            
+            markerImageCopy.innerHTML = markerImage.innerHTML;
+
             if (this.naturalSize === true) {
                 // scale to full image size
-                markerImage.width.baseVal.value = target.naturalWidth;
-                markerImage.height.baseVal.value = target.naturalHeight;
+                markerImageCopy.width.baseVal.value = target.naturalWidth;
+                markerImageCopy.height.baseVal.value = target.naturalHeight;
+            } else if (this.width !== undefined && this.height !== undefined) {
+                // scale to specific dimensions
+                markerImageCopy.width.baseVal.value = this.width;
+                markerImageCopy.height.baseVal.value = this.height;
             }
     
-            canvas.width = markerImage.width.baseVal.value;
-            canvas.height = markerImage.height.baseVal.value;
+            canvas.width = markerImageCopy.width.baseVal.value;
+            canvas.height = markerImageCopy.height.baseVal.value;
     
-            const data = markerImage.outerHTML;
-    
+            const data = markerImageCopy.outerHTML;
+
             const ctx = canvas.getContext("2d");
             if (this.markersOnly !== true) { 
                 ctx.drawImage(target, 0, 0, canvas.width, canvas.height);