+++ src/core/Settings.ts
@@ -89,5 +89,15 @@ export class Settings {
     'Courier, "Courier New", monospace',
     'cursive',
     'fantasy'
-  ]
+  ];
+
+  /**
+   * Margin in pixels between marker.js popup UI and window borders.
+   */
+  public popupMargin = 30;
+
+  /**
+   * Create a new Freehand marker for every stroke.
+   */
+  public newFreehandMarkerOnPointerUp = false;
 }