+++ src/core/Style.ts
@@ -14,10 +14,33 @@ export class Style {
   private static styleSheet?: HTMLStyleElement;
 
   /**
+   * For cases when you need to add the stylesheet to anything
+   * other than document.head (default), set this property
+   * befor calling `MarkerArea.show()`.
+   * 
+   * Example: here we set the rendering/placement root (targetRoot)
+   * to the `shadowRoot` of a web componet and set `styleSheetRoot`
+   * to the same value as well.
+   * 
+   * ```javascript
+   * const markerArea = new markerjs2.MarkerArea(target);
+   * markerArea.targetRoot = this.shadowRoot;
+   * markerjs2.Style.styleSheetRoot = this.shadowRoot;
+   * markerArea.show();
+   * ```
+   * 
+   * Known issue/limitation:
+   * you can't use marker.js 2 in both main and Shadow DOM
+   * on the same page.
+   */
+  public static styleSheetRoot: HTMLElement;
+
+  /**
    * Returns default UI styles.
    */
   public static get defaultSettings(): IStyleSettings {
     return {
+      canvasBackgroundColor: '#ffffff',
       toolbarBackgroundColor: '#111111',
       toolbarBackgroundHoverColor: '#333333',
       toolbarColor: '#eeeeee',
@@ -73,7 +96,7 @@ export class Style {
 
   private static addStyleSheet() {
     Style.styleSheet = document.createElement('style');
-    document.head.appendChild(Style.styleSheet);
+    (Style.styleSheetRoot ?? document.head).appendChild(Style.styleSheet);
 
     // add global rules
     Style.addRule(new StyleRule(`.${Style.CLASS_PREFIX} h3`, 'font-family: sans-serif'));
@@ -107,7 +130,7 @@ export class Style {
 
   public static removeStyleSheet(): void {
     if (Style.styleSheet) {
-      document.head.removeChild(Style.styleSheet);
+      (Style.styleSheetRoot ?? document.head).removeChild(Style.styleSheet);
       Style.styleSheet = undefined;
     }
   }